Add fuse_msgs pkg with Graph and Transation msgs
The Graph message uses these other new messages from fuse_msgs:
* Variable
* Constraint

For the Constraint message, there are no specific fields because they
depend on the constraint's virtual cost function. Instead, we provide a
generic multi-array of properties that rely on another new message:
* Property

A Property message has a name and a multi-array with its data. This
allows to represent things like a pose and a covariance, using two
properties. For instance, the covariance matrix can be represented by a
plain/flat multi-array.