Don't probe gradient checker b/c of segfault
refs CORE-14776

If we call Probe on the gradient checker the check passes successfully,
but the probe results produces at segfault when it's destroyed at the
end of the TEST scope.

This also happens if we create an dummy cost function and probe the
gradient checker on it, taking the gradient checker test from ceres
source code. This suggest either an issue in the gradient checker itself
or likely some compile or link issue.