Wrap normal delta pose 2d orientation angle error
If only the delta orientation angle is wrapped to the (-pi, pi]
interval, it could still happen that the error is outside of that
interval. Consider this example:
* orientation2[0] == eps
* orientation1[0] == pi
* b_(2) == pi - eps
where eps is a very small angle.

With previous implementation we get: 2 * eps - 2 * pi
With the change proposed here we get: 2 * eps