Move benchmarks into tests
refs CORE-15104

The benchmark targets are now only build if CATKIN_ENABLE_TESTING is ON,
which means that benchmark is now a test_depend and not a depend.

However, the benchmarks are NOT gtests, so they are built directly on
catkin build, i.e. there is no need to run make run_tests after. For
this reason, the find_package on benchmark is no longer REQUIRED. Not is
QUIET and the benchmark is built only if the benchmark package is FOUND.