Benchmark unicycle_2d state cost function (#121)
The benchmark targets are now only build if CATKIN_ENABLE_TESTING is ON,
which means that benchmark is now a test_depend and not a depend.

However, the benchmarks are NOT gtests, so they are built directly on
catkin build, i.e. there is no need to run make run_tests after. For
this reason, the find_package on benchmark is no longer REQUIRED,
but QUIET instead. The benchmark is built only if the benchmark package
is FOUND.