Scale process noise covariance
This scales the process noise covariance pose by the norm of the current
state velocity.

A new parameter `velocity_norm_min` is added, that prevents the process
noise scaling from setting the pose components to zero or a very small
value that could lead to NaN or a rank deficient Jacobian in the problem
solved, due to an ill-condition covariance for the process noise.