Fix getConstraints for lvalue iterator
By taking a reference to the lvalue iterator input, the changes in
`getConstraints` can be seen from outside. Note that we also need to
assign the output of `std::copy` to `result`.

With the move semantics we make `getConstraints` work for both rvalue
and lvalue inputs, e.g. `std::back_inserter<>()` and iterators.