Use twist covariance for differential dependent (#138)
In the `fuse_models::Odometry2D` sensor model, when `differential: true`
and `independent: false`, the relative pose covariance should NOT be
computed from the consecutive absolute pose covariance matrices because
they grow unbounded, so the resulting relative pose covariance suffers
from numerical issues.

Instead, we can use the twist covariance of the last pose to compute the
relative pose covariance, using the time difference between the
consecutive absolute poses.

The only limitation is that we cannot throttle the input topics, because
otherwise the twist covariance from the intermediate/throttled messages
is missed. We'll have to throttle inside the sensor model, by
integrating the intermediate messages.