Visualize loss (#144)
* Visualize error scaled by loss cost factor
If a constraint has a loss function, the cost is computed with and
without loss, and the quotient of the cost with loss by the cost without
loss is used to scale the error line. This is used to draw a "loss"
error line, so we can see the contribution of the loss function to the
constraint.