Purge transactions older than ignition max stamp
Consider an ignition sensor transaction `I` with minimum and maximum
involved stamps `[I_min, I_max]`.

And another transaction `T` with minimum and maximum involved stamps
`[T_min, T_max]`.

Now consider `I` represents an entire graph recorded, so we have
`I_max = I_min + l` where `l` is approximately the `lag_duration`.

We want to process `I` and the first transaction `T` after that graph
was generated, so it adds constraints to existing variables or removes
existing constraints or variables.

For this to happen we must purge any transaction `T` that has:
* `T_min < I_min`, or
* `T_max <= I_max`

Note that `T_max == I_max` for the `T` that generated the graph recorded,
represented by `I`.

This conditions should also be valid for other use cases.

In the particular case of `I`, we also want `T_max` to be exactly the
one after `I_max`. For this we need to disable the `reset` service in
the ignition sensor, so we do not lose any transaction.