Add invert_tf to Odometry2DPublisher
This allows to publish the inverse TF transform.

This can be useful to skip the odom->base_link lookupTransform() when
braodcasting map->base_link, which must be broadcasted as map->odom
because TF tree doesn't support more than a single parent per frame,
base_link in this case. This is particular relevant when
predict_to_current_time is enabled, because the lookupTransform() could
take a while, causing delays.