Catch graph update exceptions and log fatal msg
When the graph update method throws an exception the node crashes and
there is no way to know what transaction caused the exception because we
call notify after successfully updating the graph.

This catches any exception thrown by the graph update method and logs a
FATAL message that includes the current graph and the new transaction.